src/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.1
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==4.2.8
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-filter==23.5
django-oauth-toolkit==2.3.0
django-phonenumber-field==7.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-social-oauth2==2.1.3
drf-yasg==1.21.7
ecdsa==0.18.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jwcrypto==1.5.1
oauthlib==3.2.2
packaging==23.2
phonenumbers==8.13.30
Pillow==10.1.0
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
wrapt==1.16.0
